Extracting a common ancestor
There was a lot of duplication, which added to the chatter of the
application. Removing that in preference for a common ancestor.
