+++ app/forms/sipity/forms/etd/grad_school_signoff_form.rb
@@ -2,18 +2,7 @@ module Sipity
   module Forms
     module Etd
       # Responsible for submitting the final Grad School approval.
-      class GradSchoolSignoffForm < ProcessingActionForm
-        def initialize(attributes = {})
-          super
-          self.action = attributes.fetch(:processing_action_name) { default_processing_action_name }
-        end
-
-        attr_reader :action
-
-        def processing_action_name
-          action.name
-        end
-
+      class GradSchoolSignoffForm < Forms::StateAdvancingAction
         private
 
         def save(requested_by:)
@@ -23,16 +12,6 @@ module Sipity
               notification: "confirmation_of_grad_school_signoff", entity: work, acting_as: ['creating_user', 'etd_reviewer', 'advisor']
             )
           end
-        end
-
-        def enrichment_type
-          self.class.to_s.demodulize.underscore.sub(/_form\Z/i, '')
-        end
-        alias_method :default_processing_action_name, :enrichment_type
-
-        include Conversions::ConvertToProcessingAction
-        def action=(value)
-          @action = convert_to_processing_action(value, scope: to_processing_entity)
         end
       end
     end
